Explantion of lambda fucntion in python
Python supports the creation of anonymous functions (i.e. functions that are not bound to a name) at runtime, using a keyword called "lambda". Lambda is a very powerful concept that's well integrated into Python and is often used in conjunction with typical functional concepts like filter(), map() and reduce().